feat: 사이드바 UI 컴포넌트 분리
