+++ src/app/_components/sidebar/setting-button.tsx
@@ -1,0 +1,11 @@
+import SettingIcon from '@/icons/setting-icon';
+
+const SettingButton = () => {
+  return (
+    <div>
+      <SettingIcon color="gray" />
+    </div>
+  );
+};
+
+export default SettingButton;
